requirements.txt
asgiref==3.7.2
cachetools==5.3.1
certifi==2023.5.7
chardet==5.1.0
charset-normalizer==3.2.0
colorama==0.4.6
distlib==0.3.6
Django==4.2.3
django-resized==1.0.2
filelock==3.12.2
idna==3.4
packaging==23.1
Pillow==10.0.0
platformdirs==3.8.1
pluggy==1.2.0
pyproject-api==1.5.3
razorpay==1.3.0
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
tox==4.6.4
typing_extensions==4.7.1
urllib3==2.0.3
virtualenv==20.23.1
